requirements.txt
anyio==3.6.2
certifi==2022.12.7
click==8.1.3
dnspython==2.3.0
email-validator==1.3.0
fastapi==0.89.1
greenlet==2.0.1
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
item==0.0.3
items==0.6.5
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.1
nulltype==2.3.1
orjson==3.8.5
psycopg2-binary==2.9.5
pydantic==1.10.4
python-dotenv==0.21.0
python-multipart==0.0.5
pytils==0.4.1
PyYAML==6.0
rfc3986==1.5.0
selection==0.0.21
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.22.0
typing_extensions==4.4.0
ujson==5.7.0
user-agent==0.1.10
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
weblib==0.1.30
websockets==10.4
pytest==7.2.1
asyncpg==0.27.0
trio==0.22.0
alembic==1.9.2